fix: 引文抽屉复用阅读器的 CBETA 重排，消除句中断裂
上一版只合并了切块之间的块间距，没碰到真正的病根：chunk_text 里保留着 CBETA
每 ~18 字一次的硬换行（实测这一段 28 个 \n），而 HTML 会把文本节点里的 \n
折叠成空格。于是正文里到处是句中空格——「又經中言有三清 淨」「意牟尼即無
學意非意業」——连高亮的引文内部也被劈开。用户第二次反馈的就是这个。

阅读器一直跑 reflowText 把硬换行重排成段落与偈颂，抽屉却把原始文本直接塞进
DOM。把 reflowText（连同 TextSegment）从 TextReaderPage 抽到 utils/textReflow.ts，
两处共用一份，排版样式同样按 .reader-body 的规则给 .chat-citation-body 补齐。

高亮改在原始坐标上计算，再借每个 segment 自带的 offsets 映射回重排后的段落。
这样引文即使正好被硬换行劈开（本例的「即無\n學意非意業」）也能整句标出——
这一点靠 findQuoteSpan 的容错通道成立：它的标点集包含 \s，命中后再映射回原始
下标。

测试教训：第一版新用例是无效的。jsdom 的 textContent 原样保留 \n，不会像浏览器
那样折叠成空格，所以断言「不含『三清 淨』」在修复前也恒真。改为直接断言正文
不得残留裸换行，并实测绕过 reflow 时两条用例确实变红，才算真的守住了。
